backend/requirements.txt
flask
flask-socketio
pymavlink
pyyaml
